+++ pilot/pkg/model/endpointshards.go
@@ -403,46 +403,3 @@ func updateShardServiceAccount(shards *EndpointShards, serviceName string) bool 
 
 	return false
 }
-
-// TestEndpointIndexUpdater is an updater that will keep an EndpointIndex in sync. This is intended for tests only.
-type TestEndpointIndexUpdater struct {
-	Index *EndpointIndex
-	// Optional; if set, we will trigger ConfigUpdates in response to EDS updates as appropriate
-	ConfigUpdateFunc func(req *PushRequest)
-}
-
-var _ XDSUpdater = &TestEndpointIndexUpdater{}
-
-func NewEndpointIndexUpdater(ei *EndpointIndex) *TestEndpointIndexUpdater {
-	return &TestEndpointIndexUpdater{Index: ei}
-}
-
-func (f *TestEndpointIndexUpdater) ConfigUpdate(*PushRequest) {}
-
-func (f *TestEndpointIndexUpdater) EDSUpdate(shard ShardKey, serviceName string, namespace string, eps []*IstioEndpoint) {
-	pushType := f.Index.UpdateServiceEndpoints(shard, serviceName, namespace, eps)
-	if f.ConfigUpdateFunc != nil && (pushType == IncrementalPush || pushType == FullPush) {
-		// Trigger a push
-		f.ConfigUpdateFunc(&PushRequest{
-			Full:           pushType == FullPush,
-			ConfigsUpdated: sets.New(ConfigKey{Kind: kind.ServiceEntry, Name: serviceName, Namespace: namespace}),
-			Reason:         NewReasonStats(EndpointUpdate),
-		})
-	}
-}
-
-func (f *TestEndpointIndexUpdater) EDSCacheUpdate(shard ShardKey, serviceName string, namespace string, eps []*IstioEndpoint) {
-	f.Index.UpdateServiceEndpoints(shard, serviceName, namespace, eps)
-}
-
-func (f *TestEndpointIndexUpdater) SvcUpdate(shard ShardKey, hostname string, namespace string, event Event) {
-	if event == EventDelete {
-		f.Index.DeleteServiceShard(shard, hostname, namespace, false)
-	}
-}
-
-func (f *TestEndpointIndexUpdater) ProxyUpdate(_ cluster.ID, _ string) {}
-
-func (f *TestEndpointIndexUpdater) RemoveShard(shardKey ShardKey) {
-	f.Index.DeleteShard(shardKey)
-}
